M8/Leccion4/lab-sec-api/api/requirements.txt
fastapi
uvicorn[standard]
python-jose[cryptography]
pydantic
prometheus-fastapi-instrumentator
prometheus-client
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
opentelemetry-exporter-jaeger-thrift
deprecated
